clients/cli/pyproject.toml
[project]
name = "fafauth-cli"
version = "0.1.0"
description = "Drive the API by a script"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyjwt>=2.11.0",
    "requests>=2.32.5",
]
